requirements.txt
fastapi[standard]
uvicorn==0.34.0
motor==3.6.0
pydantic==2.10.3
pyobjectid==0.1.3
pytest==8.3.4
requests==2.32.3
python-dotenv==1.0.1
pytest-asyncio==0.25.0